{{cookiecutter.repo_name}}/{{cookiecutter.python_module_name}}/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
cfenv==0.5.3
sap-xssec==2.1.0
